# shared/openapi.json
{"openapi":"3.0.0","paths":{"/pets":{"get":{"operationId":"PetsController_findAll","summary":"List all pets","parameters":[{"name":"limit","required":false,"in":"query","description":"How many items to return at one time (max 100)","schema":{"type":"number"}}],"responses":{"200":{"description":"A paged array of pets","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Pet"}}}}}},"tags":["pets"]},"post":{"operationId":"PetsController_create","summary":"Create a pet","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePetDto"}}}},"responses":{"201":{"description":"Pet created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Pet"}}}}},"tags":["pets"]}},"/pets/{id}":{"get":{"operationId":"PetsController_findOne","summary":"Info for a specific pet","parameters":[{"name":"id","required":true,"in":"path","description":"The id of the pet to retrieve","schema":{"type":"number"}}],"responses":{"200":{"description":"Expected response to a valid request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Pet"}}}},"404":{"description":"Pet not found"}},"tags":["pets"]},"put":{"operationId":"PetsController_update","summary":"Update a pet","parameters":[{"name":"id","required":true,"in":"path","description":"The id of the pet to update","schema":{"type":"number"}}],"requestBody":{"required":true,"description":"Updated pet data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePetDto"}}}},"responses":{"200":{"description":"Pet updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Pet"}}}},"404":{"description":"Pet not found"}},"tags":["pets"]},"delete":{"operationId":"PetsController_remove","summary":"Delete a pet","parameters":[{"name":"id","required":true,"in":"path","description":"The id of the pet to delete","schema":{"type":"number"}}],"responses":{"204":{"description":"Pet deleted successfully"},"404":{"description":"Pet not found"}},"tags":["pets"]}}},"info":{"title":"Pet Store API","description":"Pet Store API with OpenAPI specification","version":"1.0","contact":{}},"tags":[{"name":"pets","description":""}],"servers":[],"components":{"schemas":{"Pet":{"type":"object","properties":{"id":{"type":"number","description":"The unique identifier of the pet","example":1,"minimum":1},"name":{"type":"string","description":"The name of the pet","example":"Fluffy"},"species":{"type":"string","description":"The species of the pet","example":"Dog"},"age":{"type":"number","description":"The age of the pet in years","example":3,"minimum":0},"description":{"type":"string","description":"Optional description of the pet","example":"A friendly golden retriever"}},"required":["id","name","species","age"]},"CreatePetDto":{"type":"object","properties":{"name":{"type":"string","description":"The name of the pet","example":"Fluffy"},"species":{"type":"string","description":"The species of the pet","example":"Dog"},"age":{"type":"number","description":"The age of the pet in years","example":3,"minimum":0},"description":{"type":"string","description":"Optional description of the pet","example":"A friendly golden retriever"}},"required":["name","species","age"]}}}}